requirements.txt
opencv-python
numpy
Pillow
python-doctr[torch]
pdf2image
openai
streamlit